+++ apps/ymir-sdl3/src/app/ui/views/debug/cdblock_filters_view.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <app/shared_context.hpp>
+
+namespace app::ui {
+
+class CDBlockFiltersView {
+public:
+    CDBlockFiltersView(SharedContext &context);
+
+    void Display();
+
+private:
+    SharedContext &m_context;
+    ymir::cdblock::CDBlock::Probe &m_probe;
+};
+
+} // namespace app::ui